SMARTVIT/requirements.txt
keras==2.3.1
tensorflow==2.3.0
# opencv==4.5.5
pillow==9.0.1
flask==2.0.3
numpy==1.16.6
paho-mqtt==1.6.1